+++ tests/testSword.cpp
@@ -10,7 +10,7 @@ TEST_CASE("Test de la classe sword")
 
     SUBCASE("Test constructeur")
     {
-        int sol=10;
+        double sol=10.0;
         sword s{sol};
 
         REQUIRE_EQ(s.getSolidity(),10);
